backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.69.0
pydantic==2.5.3
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
